# Optimize `sqlutils.ToSqlite3Dialect` for queries + DML/DDLs