--- a/go/vt/external/golib/sqlutils/sqlite_dialect.go
+++ b/go/vt/external/golib/sqlutils/sqlite_dialect.go
@@ -135,25 +135,30 @@ func ToSqlite3Insert(statement string) string {
 
 // ToSqlite3Dialect converts a statement to sqlite3 dialect. The statement
 // is checked in this order:
-//  1. If an insert/replace, convert with ToSqlite3Insert.
-//  2. If a create table, convert with IsCreateTable.
-//  3. If an alter table, convert with IsAlterTable.
-//  4. As fallback, return the statement with sqlite3GeneralConversions applied.
-func ToSqlite3Dialect(statement string) (translated string) {
-	if IsInsert(statement) {
-		return ToSqlite3Insert(statement)
-	}
-	if IsCreateIndex(statement) {
-		return ToSqlite3CreateIndex(statement)
-	}
-	if IsDropIndex(statement) {
-		return ToSqlite3DropIndex(statement)
-	}
-	if IsCreateTable(statement) {
-		return ToSqlite3CreateTable(statement)
-	}
-	if IsAlterTable(statement) {
-		return ToSqlite3CreateTable(statement)
+//  1. If a query, return the statement with sqlite3GeneralConversions applied.
+//  2. If an insert/replace, convert with ToSqlite3Insert.
+//  3. If a create index, convert with IsCreateIndex.
+//  4. If an drop table, convert with IsDropIndex.
+//  5. If a create table, convert with IsCreateTable.
+//  6. If an alter table, convert with IsAlterTable.
+//  7. As fallback, return the statement with sqlite3GeneralConversions applied.
+func ToSqlite3Dialect(statement string, potentiallyDMLOrDDL bool) (translated string) {
+	if potentiallyDMLOrDDL {
+		if IsInsert(statement) {
+			return ToSqlite3Insert(statement)
+		}
+		if IsCreateIndex(statement) {
+			return ToSqlite3CreateIndex(statement)
+		}
+		if IsDropIndex(statement) {
+			return ToSqlite3DropIndex(statement)
+		}
+		if IsCreateTable(statement) {
+			return ToSqlite3CreateTable(statement)
+		}
+		if IsAlterTable(statement) {
+			return ToSqlite3CreateTable(statement)
+		}
 	}
 	return applyConversions(statement, sqlite3GeneralConversions)
 }