## Description

This PR adds some slight optimisations to `sqlutils.ToSqlite3Dialect`, a function we've noticed to take up a significant amount of CPU time in `vtorc` when it watches many instances, primarily due to large insert queries needing to be parsed by `regexp` filters

Profile SVG _(need to enlarge)_:
![profile001](https://github.com/user-attachments/assets/60a719c2-ef26-4011-a18f-1797cf9c980e)

All this change does is assumes inserts happen more frequently than alter/create. Moving inserts to be checked sooner and separating read queries causes them to have to go through [fewer regex matches](https://github.com/vitessio/vitess/blob/main/go/vt/external/golib/sqlutils/sqlite_dialect.go) _(still a lot 🙈)_. The expense of queries that are not insert, alter or create is unchanged.

On Slack's `vtorc` instances this gave us about a 15-20% reduction in user-CPU used by `vtorc` vs the previous avg

In the benchmark I added in the PR the difference is more like 8-9% for big inserts and 18-19% for reads

#### Before
```bash
goos: darwin
goarch: amd64
pkg: vitess.io/vitess/go/vt/external/golib/sqlutils
cpu: Intel(R) Core(TM) i9-9980HK CPU @ 2.40GHz
BenchmarkToSqlite3Dialect_Insert1000
BenchmarkToSqlite3Dialect_Insert1000-16    	       1	1870643229 ns/op
BenchmarkToSqlite3Dialect_Select
BenchmarkToSqlite3Dialect_Select-16        	    3541	    330667 ns/op
PASS
ok  	vitess.io/vitess/go/vt/external/golib/sqlutils	4.660s
```

#### After
```bash
goos: darwin
goarch: amd64
pkg: vitess.io/vitess/go/vt/external/golib/sqlutils
cpu: Intel(R) Core(TM) i9-9980HK CPU @ 2.40GHz
BenchmarkToSqlite3Dialect_Insert1000
BenchmarkToSqlite3Dialect_Insert1000-16    	       1	1711349181 ns/op
BenchmarkToSqlite3Dialect_Select
BenchmarkToSqlite3Dialect_Select-16        	    4436	    270223 ns/op
PASS
ok  	vitess.io/vitess/go/vt/external/golib/sqlutils	4.528s
```

## Related Issue(s)

Resolves https://github.com/vitessio/vitess/issues/17114

## Checklist

-   [x] "Backport to:" labels have been added if this change should be back-ported to release branches
-   [x] If this change is to be back-ported to previous releases, a justification is included in the PR description
-   [x] Tests were added or are not required
-   [x] Did the new or modified tests pass consistently locally and on CI?
-   [x] Documentation was added or is not required

## Deployment Notes

<!-- Notes regarding deployment of the contained body of work. These should note any db migrations, etc. -->