@mattlord 2024-10-29
@GuptaManan100 do we still need this library at all? `vtorc` only supports SQLite now so I would think that we would use the SQLite dialect everywhere.