@timvaillancourt 2024-10-29
> @GuptaManan100 do we still need this library at all? `vtorc` only supports SQLite now so I would think that we would use the SQLite dialect everywhere.

@mattlord his is what I asked/proposed [in this thread](https://vitess.slack.com/archives/C02GSRZ8XAN/p1730207109838589) 👍

I'm hoping we can remove dialect translation and just use sqlite dialect. Also I think `external/sqlutils` could be replaced with something a bit less crufty - there are similar projects that can reflect `database/sql` to map/structs for example. For now this is some duct tape