@dbussink 2024-10-29
> @mattlord this is what I asked/proposed [in this thread](https://vitess.slack.com/archives/C02GSRZ8XAN/p1730207109838589) 👍. I'm hoping we can remove dialect translation and just use sqlite dialect

I think removing the translation and writing queries directly that work with sqlite is a good idea indeed. It removes this whole layer of translation that doesn't really serve any purpose anymore.

> Also I think `external/sqlutils` could be replaced with something a bit less crufty - there are similar projects that can reflect `database/sql` to map/structs for example. For now this is some duct tape

I think it's a good idea to clean this up (also fine incrementally). It's all leftovers from how it was basically vendored but only parts have been modernized to match how other parts of Vitess work. So I think thinks kind of cleanup is also very welcome.