@GuptaManan100 2024-10-30
Yes, @timvaillancourt we can get rid of all other dialects other than sqlite. We don't allow any other backing database in VTOrc.