@timvaillancourt 2024-10-31
Closing in favour of https://github.com/vitessio/vitess/pull/17124 which removes the translation entirely